Implement simple MessageBox dialog, where no icon will be displayed.
In some cases it was necessary to have a dialog without the icon. Especially on macOS where the dialog's text gets squeezed due to the icon being too big, making it look a little odd.
On Linux I made it display as `info`, since I don't know of a way to disable icons using zenity.